requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-community>=0.0.13
langchain-classic>=1.0.0
tavily-python>=0.3.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
PyJWT>=2.8.0
crewai>=0.80.0
crewai-tools>=0.12.0
langsmith>=0.1.0
colorama>=0.4.6
textstat>=0.7.3
beautifulsoup4>=4.12.0
jupyter>=1.0.0
ipykernel>=6.0.0
